adds remaining controllers
