+++ src/views/languages.php
@@ -9,51 +9,14 @@
 // Include application functions
 require_once '../lib.inc.php';
 
-$action = (isset($_REQUEST['action'])) ? $_REQUEST['action'] : '';
-if (!isset($msg)) {
-	$msg = '';
-}
-
-/**
- * Show default list of languages in the database
- */
-function doDefault($msg = '') {
-	global $data, $misc, $database;
-	global $lang;
-
-	$misc->printTrail('database');
-	$misc->printTabs('database', 'languages');
-	$misc->printMsg($msg);
-
-	$languages = $data->getLanguages();
-
-	$columns = [
-		'language' => [
-			'title' => $lang['strname'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('lanname'),
-		],
-		'trusted' => [
-			'title' => $lang['strtrusted'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('lanpltrusted'),
-			'type' => 'yesno',
-		],
-		'function' => [
-			'title' => $lang['strfunction'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('lanplcallf'),
-		],
-	];
-
-	$actions = [];
-
-	echo $misc->printTable($languages, $columns, $actions, 'languages-languages', $lang['strnolanguages']);
-}
+$lang_controller = new \PHPPgAdmin\Controller\LangController($container);
 
 $misc->printHeader($lang['strlanguages']);
 $misc->printBody();
 
 switch ($action) {
 	default:
-		doDefault();
+		$lang_controller->doDefault();
 		break;
 }
