+++ src/views/operators.php
@@ -9,181 +9,7 @@
 // Include application functions
 require_once '../lib.inc.php';
 
-$action = (isset($_REQUEST['action'])) ? $_REQUEST['action'] : '';
-if (!isset($msg)) {
-	$msg = '';
-}
-
-/**
- * Show read only properties for an operator
- */
-function doProperties($msg = '') {
-	global $data, $misc;
-	global $lang;
-
-	$misc->printTrail('operator');
-	$misc->printTitle($lang['strproperties'], 'pg.operator');
-	$misc->printMsg($msg);
-
-	$oprdata                       = $data->getOperator($_REQUEST['operator_oid']);
-	$oprdata->fields['oprcanhash'] = $data->phpBool($oprdata->fields['oprcanhash']);
-
-	if ($oprdata->recordCount() > 0) {
-		echo "<table>\n";
-		echo "<tr><th class=\"data left\">{$lang['strname']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprname']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strleftarg']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprleftname']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strrightarg']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprrightname']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strcommutator']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprcom']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strnegator']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprnegate']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strjoin']}</th>\n";
-		echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprjoin']), "</td></tr>\n";
-		echo "<tr><th class=\"data left\">{$lang['strhashes']}</th>\n";
-		echo "<td class=\"data1\">", ($oprdata->fields['oprcanhash']) ? $lang['stryes'] : $lang['strno'], "</td></tr>\n";
-
-		/* these field only exists in 8.2 and before in pg_catalog */
-		if (isset($oprdata->fields['oprlsortop'])) {
-			echo "<tr><th class=\"data left\">{$lang['strmerges']}</th>\n";
-			echo "<td class=\"data1\">", ($oprdata->fields['oprlsortop'] !== '0' && $oprdata->fields['oprrsortop'] !== '0') ? $lang['stryes'] : $lang['strno'], "</td></tr>\n";
-			echo "<tr><th class=\"data left\">{$lang['strrestrict']}</th>\n";
-			echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprrest']), "</td></tr>\n";
-			echo "<tr><th class=\"data left\">{$lang['strleftsort']}</th>\n";
-			echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprlsortop']), "</td></tr>\n";
-			echo "<tr><th class=\"data left\">{$lang['strrightsort']}</th>\n";
-			echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprrsortop']), "</td></tr>\n";
-			echo "<tr><th class=\"data left\">{$lang['strlessthan']}</th>\n";
-			echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprltcmpop']), "</td></tr>\n";
-			echo "<tr><th class=\"data left\">{$lang['strgreaterthan']}</th>\n";
-			echo "<td class=\"data1\">", $misc->printVal($oprdata->fields['oprgtcmpop']), "</td></tr>\n";
-		} else {
-			echo "<tr><th class=\"data left\">{$lang['strmerges']}</th>\n";
-			echo "<td class=\"data1\">", $data->phpBool($oprdata->fields['oprcanmerge']) ? $lang['stryes'] : $lang['strno'], "</td></tr>\n";
-		}
-		echo "</table>\n";
-
-		$misc->printNavLinks([
-			'showall' => [
-				'attr' => [
-					'href' => [
-						'url' => 'operators.php',
-						'urlvars' => [
-							'server' => $_REQUEST['server'],
-							'database' => $_REQUEST['database'],
-							'schema' => $_REQUEST['schema'],
-						],
-					],
-				],
-				'content' => $lang['strshowalloperators'],
-			]], 'operators-properties', get_defined_vars()
-		);
-	} else {
-		doDefault($lang['strinvalidparam']);
-	}
-
-}
-
-/**
- * Show confirmation of drop and perform actual drop
- */
-function doDrop($confirm) {
-	global $data, $misc;
-	global $lang;
-
-	if ($confirm) {
-		$misc->printTrail('operator');
-		$misc->printTitle($lang['strdrop'], 'pg.operator.drop');
-
-		echo "<p>", sprintf($lang['strconfdropoperator'], $misc->printVal($_REQUEST['operator'])), "</p>\n";
-
-		echo "<form action=\"/src/views/operators.php\" method=\"post\">\n";
-		echo "<p><input type=\"checkbox\" id=\"cascade\" name=\"cascade\" /> <label for=\"cascade\">{$lang['strcascade']}</label></p>\n";
-		echo "<p><input type=\"hidden\" name=\"action\" value=\"drop\" />\n";
-		echo "<input type=\"hidden\" name=\"operator\" value=\"", htmlspecialchars($_REQUEST['operator']), "\" />\n";
-		echo "<input type=\"hidden\" name=\"operator_oid\" value=\"", htmlspecialchars($_REQUEST['operator_oid']), "\" />\n";
-		echo $misc->form;
-		echo "<input type=\"submit\" name=\"drop\" value=\"{$lang['strdrop']}\" />\n";
-		echo "<input type=\"submit\" name=\"cancel\" value=\"{$lang['strcancel']}\" /></p>\n";
-		echo "</form>\n";
-	} else {
-		$status = $data->dropOperator($_POST['operator_oid'], isset($_POST['cascade']));
-		if ($status == 0) {
-			doDefault($lang['stroperatordropped']);
-		} else {
-			doDefault($lang['stroperatordroppedbad']);
-		}
-
-	}
-
-}
-
-/**
- * Show default list of operators in the database
- */
-function doDefault($msg = '') {
-	global $data, $conf, $misc;
-	global $lang;
-
-	$misc->printTrail('schema');
-	$misc->printTabs('schema', 'operators');
-	$misc->printMsg($msg);
-
-	$operators = $data->getOperators();
-
-	$columns = [
-		'operator' => [
-			'title' => $lang['stroperator'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('oprname'),
-			'url' => "operators.php?action=properties&amp;{$misc->href}&amp;",
-			'vars' => ['operator' => 'oprname', 'operator_oid' => 'oid'],
-		],
-		'leftarg' => [
-			'title' => $lang['strleftarg'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('oprleftname'),
-		],
-		'rightarg' => [
-			'title' => $lang['strrightarg'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('oprrightname'),
-		],
-		'returns' => [
-			'title' => $lang['strreturns'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('resultname'),
-		],
-		'actions' => [
-			'title' => $lang['stractions'],
-		],
-		'comment' => [
-			'title' => $lang['strcomment'],
-			'field' => \PHPPgAdmin\Decorators\Decorator::field('oprcomment'),
-		],
-	];
-
-	$actions = [
-		'drop' => [
-			// 'title' => $lang['strdrop'],
-			// 'url'   => "operators.php?action=confirm_drop&amp;{$misc->href}&amp;",
-			// 'vars'  => array('operator' => 'oprname', 'operator_oid' => 'oid'),
-			'content' => $lang['strdrop'],
-			'attr' => [
-				'href' => [
-					'url' => 'operators.php',
-					'urlvars' => [
-						'action' => 'confirm_drop',
-						'operator' => \PHPPgAdmin\Decorators\Decorator::field('oprname'),
-						'operator_oid' => \PHPPgAdmin\Decorators\Decorator::field('oid'),
-					],
-				],
-			],
-		],
-	];
-
-	echo $misc->printTable($operators, $columns, $actions, 'operators-operators', $lang['strnooperators']);
-
-//		TODO operators.php action=create $lang['strcreateoperator']
-}
+$operator_controller = new \PHPPgAdmin\Controller\OperatorController($container);
 
 $misc->printHeader($lang['stroperators']);
 $misc->printBody();
@@ -191,9 +17,9 @@ $misc->printBody();
 switch ($action) {
 	case 'save_create':
 		if (isset($_POST['cancel'])) {
-			doDefault();
+			$operator_controller->doDefault();
 		} else {
-			doSaveCreate();
+			$operator_controller->doSaveCreate();
 		}
 
 		break;
@@ -202,20 +28,20 @@ switch ($action) {
 		break;
 	case 'drop':
 		if (isset($_POST['cancel'])) {
-			doDefault();
+			$operator_controller->doDefault();
 		} else {
-			doDrop(false);
+			$operator_controller->doDrop(false);
 		}
 
 		break;
 	case 'confirm_drop':
-		doDrop(true);
+		$operator_controller->doDrop(true);
 		break;
 	case 'properties':
-		doProperties();
+		$operator_controller->doProperties();
 		break;
 	default:
-		doDefault();
+		$operator_controller->doDefault();
 		break;
 }
